# assets/openapi.json
{
    "openapi": "3.0.0",
    "servers": [
        {
            "url": "http://localhost:3001/api/",
            "description": "Official Solarcord Instance"
        }
    ],
    "info": {
        "description": "Solarcord is a free open source selfhostable discord compatible chat, voice and video platform",
        "version": "1.0.0",
        "title": "Solarcord HTTP API Routes",
        "termsOfService": "",
        "contact": {
            "name": "Solarcord"
        },
        "license": {
            "name": "AGPLV3",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer",
                "description": "Bearer/Bot prefixes are not required."
            }
        },
        "schemas": {
...
            "ChannelPermissionOverwriteType": {
                "enum": [
                    0,
                    1,
                    2
                ],
                "type": "number"
            },
...
            "ChannelModifySchema": {
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 100,
                        "type": "string"
                    },
                    "type": {
                        "enum": [
                            0,
                            1,
                            10,
                            11,
                            12,
                            13,
                            14,
                            15,
                            2,
                            255,
                            3,
                            33,
                            34,
                            35,
                            4,
                            5,
                            6,
                            64,
                            7,
                            8,
                            9
                        ],
                        "type": "number"
                    },
                    "topic": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string",
                        "nullable": true
                    },
                    "bitrate": {
                        "type": "integer"
                    },
                    "user_limit": {
                        "type": "integer"
                    },
                    "rate_limit_per_user": {
                        "type": "integer"
                    },
                    "position": {
                        "type": "integer"
                    },
                    "permission_overwrites": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "type": {
                                    "$ref": "#/components/schemas/ChannelPermissionOverwriteType"
                                },
                                "allow": {
                                    "type": "string"
                                },
                                "deny": {
                                    "type": "string"
                                }
                            },
                            "additionalProperties": false,
                            "required": [
                                "allow",
                                "deny",
                                "id",
                                "type"
                            ]
                        }
                    },
                    "parent_id": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "nsfw": {
                        "type": "boolean"
                    },
                    "rtc_region": {
                        "type": "string"
                    },
                    "default_auto_archive_duration": {
                        "type": "integer"
                    },
                    "default_reaction_emoji": {
                        "type": "string",
                        "nullable": true
                    },
                    "flags": {
                        "type": "integer"
                    },
                    "default_thread_rate_limit_per_user": {
                        "type": "integer"
                    },
                    "video_quality_mode": {
                        "type": "integer"
                    }
                }
            },
...
            "GuildCreateSchema": {
                "type": "object",
                "properties": {
                    "name": {
                        "maxLength": 100,
                        "type": "string"
                    },
                    "region": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string",
                        "nullable": true
                    },
                    "channels": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ChannelModifySchema"
                        }
                    },
                    "guild_template_code": {
                        "type": "string"
                    },
                    "system_channel_id": {
                        "type": "string"
                    },
                    "rules_channel_id": {
                        "type": "string"
                    }
                }
            },
...
            "GuildUpdateWelcomeScreenSchema": {
                "type": "object",
                "properties": {
                    "welcome_channels": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "channel_id": {
                                    "type": "string"
                                },
                                "description": {
                                    "type": "string"
                                },
                                "emoji_id": {
                                    "type": "string"
                                },
                                "emoji_name": {
                                    "type": "string"
                                }
                            },
                            "additionalProperties": false,
                            "required": [
                                "channel_id",
                                "description"
                            ]
                        }
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "description": {
                        "type": "string"
                    }
                }
            },
...
            "MemberChangeSchema": {
                "type": "object",
                "properties": {
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "nick": {
                        "type": "string"
                    },
                    "avatar": {
                        "type": "string",
                        "nullable": true
                    },
                    "bio": {
                        "type": "string"
                    }
                }
            },
...
            "DmChannelCreateSchema": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "recipients": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "recipients"
                ]
            },
...
            "GuildUpdateSchema": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "banner": {
                        "type": "string",
                        "nullable": true
                    },
                    "splash": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string"
                    },
                    "features": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "verification_level": {
                        "type": "integer"
                    },
                    "default_message_notifications": {
                        "type": "integer"
                    },
                    "system_channel_flags": {
                        "type": "integer"
                    },
                    "explicit_content_filter": {
                        "type": "integer"
                    },
                    "public_updates_channel_id": {
                        "type": "string"
                    },
                    "afk_timeout": {
                        "type": "integer"
                    },
                    "afk_channel_id": {
                        "type": "string"
                    },
                    "preferred_locale": {
                        "type": "string"
                    },
                    "premium_progress_bar_enabled": {
                        "type": "boolean"
                    },
                    "discovery_splash": {
                        "type": "string"
                    },
                    "region": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string",
                        "nullable": true
                    },
                    "guild_template_code": {
                        "type": "string"
                    },
                    "system_channel_id": {
                        "type": "string"
                    },
                    "rules_channel_id": {
                        "type": "string"
                    }
                }
            },
...
            "ChannelReorderSchema": {
                "type": "array",
                "items": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string"
                        },
                        "position": {
                            "type": "integer"
                        },
                        "lock_permissions": {
                            "type": "boolean"
                        },
                        "parent_id": {
                            "type": "string"
                        }
                    },
                    "additionalProperties": false,
                    "required": [
                        "id"
                    ]
                }
            },
...
        }
    },
...
    "paths": {
...
        "/users/@me/guilds/": {
            "get": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "tags": [
                    "users"
                ]
            }
        },
        "/users/@me/guilds/{guild_id}": {
            "delete": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "guild_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "guild_id"
                    }
                ],
                "tags": [
                    "users"
                ]
            }
        },
...
        "/users/@me/channels/": {
            "get": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "tags": [
                    "users"
                ]
            },
            "post": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DmChannelCreateSchema"
                            }
                        }
                    }
                },
                "tags": [
                    "users"
                ]
            }
        },
...
        "/users/@me/affinities/guilds/": {
            "get": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "tags": [
                    "users"
                ]
            }
        },
...
        "/guilds/": {
            "post": {
                "x-right-required": "CREATE_GUILDS",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GuildCreateSchema"
                            }
                        }
                    }
                },
                "tags": [
                    "guilds"
                ]
            }
        },
...
        "/guilds/{guild_id}/welcome-screen/": {
            "get": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "guild_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "guild_id"
                    }
                ],
                "tags": [
                    "guilds"
                ]
            },
            "patch": {
                "x-permission-required": "MANAGE_GUILD",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GuildUpdateWelcomeScreenSchema"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "guild_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "guild_id"
                    }
                ],
                "tags": [
                    "guilds"
                ]
            }
        },
...
        "/guilds/{guild_id}/members/": {
            "get": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "guild_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "guild_id"
                    }
                ],
                "tags": [
                    "guilds"
                ]
            }
        },
...
        "/guilds/{guild_id}/members/{member_id}/": {
            "get": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "guild_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "guild_id"
                    },
                    {
                        "name": "member_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "member_id"
                    }
                ],
                "tags": [
                    "guilds"
                ]
            },
            "patch": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MemberChangeSchema"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "guild_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "guild_id"
                    },
                    {
                        "name": "member_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "member_id"
                    }
                ],
                "tags": [
                    "guilds"
                ]
            },
            "put": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "guild_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "guild_id"
                    },
                    {
                        "name": "member_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "member_id"
                    }
                ],
                "tags": [
                    "guilds"
                ]
            },
            "delete": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "guild_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "guild_id"
                    },
                    {
                        "name": "member_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "member_id"
                    }
                ],
                "tags": [
                    "guilds"
                ]
            }
        },
...
        "/guilds/{guild_id}/": {
            "get": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "guild_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "guild_id"
                    }
                ],
                "tags": [
                    "guilds"
                ]
            },
            "patch": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GuildUpdateSchema"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "guild_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "guild_id"
                    }
                ],
                "tags": [
                    "guilds"
                ]
            }
        },
...
        "/guilds/{guild_id}/channels/": {
            "get": {
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "guild_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "guild_id"
                    }
                ],
                "tags": [
                    "guilds"
                ]
            },
            "post": {
                "x-permission-required": "MANAGE_CHANNELS",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChannelModifySchema"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "guild_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "guild_id"
                    }
                ],
                "tags": [
                    "guilds"
                ]
            },
            "patch": {
                "x-permission-required": "MANAGE_CHANNELS",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChannelReorderSchema"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "guild_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "guild_id"
                    }
                ],
                "tags": [
                    "guilds"
                ]
            }
        },
...
        "/channels/{channel_id}/": {
            "get": {
                "x-permission-required": "VIEW_CHANNEL",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            },
            "delete": {
                "x-permission-required": "MANAGE_CHANNELS",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            },
            "patch": {
                "x-permission-required": "MANAGE_CHANNELS",
                "security": [
                    {
                        "bearer": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ChannelModifySchema"
                            }
                        }
                    }
                },
                "parameters": [
                    {
                        "name": "channel_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "channel_id"
                    }
                ],
                "tags": [
                    "channels"
                ]
            }
        },
...
    }
}